Ajustado Removacao de OLT

+++ classes/deletar.php
@@ -1,5 +1,6 @@
 <?php
   include_once "../db/db_config_mysql.php";
+  include_once "../db/db_config_radius.php";
   // Inicia sessões 
   session_start();
 
@@ -11,34 +12,51 @@
       $contrato = $_POST["contrato"];
       $serial = $_POST["serial"];
 
-      $sql_disponibiliza_porta = "UPDATE ctos SET porta_atendimento_disponivel = 0 WHERE serial = '$serial'";
+      
+      $sql_disponibiliza_porta = "UPDATE ctos SET porta_atendimento_disponivel = 1 WHERE serial = '$serial'";
       $executa_query = mysqli_query($conectar,$sql_disponibiliza_porta);
 
       if ($executa_query) 
       {
-        $sql_apagar_onu = ("DELETE FROM ont WHERE contrato = '$contrato' AND serial = '$serial'" );
+        $deletar_onu_radius_banda = "DELETE FROM radreply WHERE username='2500/13/0/$serial@vertv' 
+          AND attribute='Huawei-Qos-Profile-Name' ";
+        $executa_query= mysqli_query($conectar_radius,$deletar_onu_radius_banda);
 
-        $deletar_onu = mysqli_query($conectar,$sql_apagar_onu);
-        if($deletar_onu)
+        $deletar_onu_radius = " DELETE FROM radcheck WHERE username='2500/13/0/$serial@vertv' ";
+        $executa_query_radius = mysqli_query($conectar_radius,$deletar_onu_radius);
+
+        if ($executa_query && $deletar_onu_radius) 
         {
-            if ( $total = mysqli_affected_rows($conectar))    //retorna quantas rows foram afetadas           
-            {
-                $_SESSION['menssagem'] = "$total ONU Removida!";
-                header('Location: ../ont_classes/ont_delete.php');
-                mysqli_close($conectar);
-                exit;
-            }else{
-                $_SESSION['menssagem'] = "ONU Não Removida!";
-                header('Location: ../ont_classes/ont_delete.php');
-                mysqli_close($conectar);
-                exit;
-            }
+          $sql_apagar_onu = ("DELETE FROM ont WHERE contrato = '$contrato' AND serial = '$serial'" );
+
+          $deletar_onu = mysqli_query($conectar,$sql_apagar_onu);
+          if($deletar_onu)
+          {
+              if ( $total = mysqli_affected_rows($conectar))    //retorna quantas rows foram afetadas           
+              {
+                  $_SESSION['menssagem'] = "$total ONU Removida!";
+                  header('Location: ../ont_classes/ont_delete.php');
+                  mysqli_close($conectar);
+                  exit;
+              }else{
+                  $_SESSION['menssagem'] = "ONU Não Removida!";
+                  header('Location: ../ont_classes/ont_delete.php');
+                  mysqli_close($conectar);
+                  exit;
+              }
+          }else{
+              $erro = mysqli_error($conectar);
+              $_SESSION['menssagem'] = "Houve erro ao deletar SQL: $erro";
+              header('Location: ../ont_classes/ont_delete.php');
+              mysqli_close($conectar);
+              exit;
+          }
         }else{
-            $erro = mysqli_error($conectar);
-            $_SESSION['menssagem'] = "Houve erro ao deletar SQL: $erro";
-            header('Location: ../ont_classes/ont_delete.php');
-            mysqli_close($conectar);
-            exit;
+          $erro = mysqli_error($conectar);
+          $_SESSION['menssagem'] = "Houve erro ao deletar SQL Radius: $erro";
+          header('Location: ../ont_classes/ont_delete.php');
+          mysqli_close($conectar);
+          exit;
         }
       }else{
         $erro = mysqli_error($conectar);
